Switch from uint64_t to size_t for return type

+++ kernel/include/cernel/lib/memory.h
@@ -2,4 +2,4 @@
 #include <stdint.h>
 #include <stivale/stivale.h>
 
-uint64_t get_memory_size(struct stivale_mmap_entry *memory_map_addr, uint32_t count);
+size_t get_memory_size(struct stivale_mmap_entry *memory_map_addr, uint32_t count);
